+++ src/components/templates/MultistepFormPage/DataProvider.tsx
@@ -1,7 +1,6 @@
 import React, { useState } from 'react';
 import { StepKey } from '@site/src/components/templates/MultistepFormPage/types';
-
-const BASE_URL = 'multistep-form-page';
+import { getActiveStep, getStepUrl } from '@site/src/components/templates/MultistepFormPage/utils/steps';
 
 interface DataProviderProps {
   children: React.ReactElement;
@@ -12,15 +11,6 @@ export const formSteps = [
   { id: StepKey.Step2, name: 'Additional information' },
   { id: StepKey.Step3, name: 'Preview' },
 ];
-
-export const getStepUrl = (step: StepKey) => {
-  return `${BASE_URL}#${step}`;
-};
-
-const getActiveStep = () => {
-  const step = window.location.hash || '#step1';
-  return step.replace('#', '');
-};
 
 /**
  * A wrapper component to provide data and action props to the child components .
